moving types to a different file
